requirements.txt
Django==1.7
dj-database-url==0.3.0
dj-static==0.0.6
django-autoslug==1.7.2
gunicorn==19.1.1
requests==2.4.1
static3==0.5.1
wsgiref==0.1.2
